feat: bias 범위 재수정

+++ topicurator/src/main/java/com/odte/topicurator/proscons/infrastructure/LlmSummarizer.java
@@ -1,6 +1,6 @@
 package com.odte.topicurator.proscons.infrastructure;
 
 public interface LlmSummarizer {
-    record Result(String title, String summary, String pros, String neutral, String cons, int bias) {}
+    record Result(String title, String summary, String pros, String neutral, String cons, Short bias) {}
     Result summarizeFromUrl(String url);
 }
